pyproject.toml
[project]
name = "expensesyncbot"
version = "0.2.0"
description = "Automated expense tracking with MCP integration and OpenAI Agents"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Aitor" }
]

dependencies = [
    "openai-agents>=0.0.3",
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
